+++ include/boost/http_proto.hpp
@@ -18,8 +18,6 @@
 #include <boost/http_proto/field.hpp>
 #include <boost/http_proto/fields.hpp>
 #include <boost/http_proto/fields_view.hpp>
-#include <boost/http_proto/headers.hpp>
-#include <boost/http_proto/headers_view.hpp>
 #include <boost/http_proto/method.hpp>
 #include <boost/http_proto/parser.hpp>
 #include <boost/http_proto/request.hpp>
