--- include/boost/http_proto/impl/headers_view.ipp
@@ -1,21 +1,0 @@
-//
-// Copyright (c) 2016-2019 Vinnie Falco (vinnie dot falco at gmail dot com)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-// Official repository: https://github.com/vinniefalco/http_proto
-//
-
-#ifndef BOOST_HTTP_PROTO_IMPL_HEADERS_VIEW_IPP
-#define BOOST_HTTP_PROTO_IMPL_HEADERS_VIEW_IPP
-
-#include <boost/http_proto/headers_view.hpp>
-
-namespace boost {
-namespace http_proto {
-
-} // http_proto
-} // boost
-
-#endif
