+++ include/boost/http_proto/src.hpp
@@ -35,8 +35,6 @@ in a translation unit of the program.
 #include <boost/http_proto/impl/field.ipp>
 #include <boost/http_proto/impl/fields.ipp>
 #include <boost/http_proto/impl/fields_view.ipp>
-#include <boost/http_proto/impl/headers.ipp>
-#include <boost/http_proto/impl/headers_view.ipp>
 #include <boost/http_proto/impl/method.ipp>
 #include <boost/http_proto/impl/parser.ipp>
 #include <boost/http_proto/impl/request.ipp>
