--- test/headers.cpp
@@ -1,11 +1,0 @@
-//
-// Copyright (c) 2019 Vinnie Falco (vinnie.falco@gmail.com)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-// Official repository: https://github.com/vinniefalco/http_proto
-//
-
-// Test that header file is self-contained.
-#include <boost/http_proto/headers.hpp>
